Fixed linting errors, which required renaming signalling to signaling
