+++ signaling/message.go
@@ -1,4 +1,4 @@
-package signalling
+package signaling
 
 import (
 	"encoding/json"
