+++ signaling/peer.go
@@ -1,4 +1,4 @@
-package signalling
+package signaling
 
 import (
 	"sync"
